pyproject.toml
[tool.poetry]
name = "pythonb_practice"
version = "0.1.0"
description = ""
authors = ["Aditya Gudimella <aditya.gudimella@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
hypothesis = "^5.1.0"
pyhamcrest = "^1.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
